Extract conformance logic to lib and tighten external-edge metric
Two changes to the reference conformance report:

The graph build, file enumeration, component analysis, and compute_report function now live in lib/conformance.py. The entry point script keeps argument parsing, the human-readable formatter, and exit-status logic only — matching the thin-wrapper rule for entry points under skill-system-foundry/scripts/. compute_report stays re-exported from the entry point so existing callers keep working.

external_edges_per_capability previously incremented for any capability-scope link whose normalized path started with ../. That over-counted intra-capability traversals like ../capability.md from capabilities/<n>/references/foo.md, which stay within the capability scope and are not lift-rewrite candidates. The metric now requires the resolved path to escape the capability root via is_within_directory, matching the documented definition ("edges into the shared skill root").

The top-level docstring is also corrected — the previous "all .md files in the skill" wording omitted the scripts/ and assets/ exclusion the walker enforces. The metrics list is updated to describe the external-edge filter accurately.

Adds a regression test for the intra-capability ../capability.md case.
